refactor(plugins): split builtin templates into folders
